ask before removing minion-master pub key

+++ configure_machine/bootstrap_bevy_member_here.py
@@ -598,28 +598,32 @@ if __name__ == '__main__':
         print('You may continue to use that master, and add a second master for your bevy.')
         run_second_minion = affirmative(input('Do you wish to run a second minion? [y/N]:'))
     two = '2' if run_second_minion else ''
-    if platform.system() == 'Windows':
-        master_pub = Path(r'C:\salt{}\conf\pki\minion\minion_master.pub'.format(two))
-    else:
-        master_pub = Path('/etc/salt{}/pki/minion/minion_master.pub'.format(two))
-    try:  # forget a former master's key (if any)
-        print('Removing master public key "{}"'.format(master_pub))
-        master_pub.unlink()
-    except (FileNotFoundError, PermissionError):
-        pass
 
     # if there is no top.sls, copy ours to make a start
-    if not os.path.exists(SALT_SRV_ROOT + 'top.sls'):
+    if not os.path.exists(SALT_SRV_ROOT + '/top.sls'):
         print('Creating a new default {}/top.sls'.format(SALT_SRV_ROOT))
         os.makedirs(SALT_SRV_ROOT, exist_ok=True)  # 3.4
         shutil.copy('../bevy_srv/salt/top.sls', SALT_SRV_ROOT)
-    if not os.path.exists(SALT_PILLAR_ROOT + 'top.sls'):
+    if not os.path.exists(SALT_PILLAR_ROOT + '/top.sls'):
         print('Creating a new default {}/top.sls'.format(SALT_PILLAR_ROOT))
         os.makedirs(SALT_PILLAR_ROOT, exist_ok=True)  # 3.4
         shutil.copy('../bevy_srv/pillar/top.sls', SALT_PILLAR_ROOT)
 
     master_address = choose_master_address(settings.get('bevymaster_url', master_id))
     settings['bevymaster_url'] = master_address
+
+    if platform.system() == 'Windows':
+        master_pub = Path(r'C:\salt{}\conf\pki\minion\minion_master.pub'.format(two))
+    else:
+        master_pub = Path('/etc/salt{}/pki/minion/minion_master.pub'.format(two))
+    if os.path.exists(master_pub):
+        if affirmative(input('Will this be a new minion<-->master relationship? [y/N]:')):
+            print("** Remember to accept this machine's Minion key on its new Master. **")
+            try:  # forget a former master's key (if any)
+                print('Removing master public key "{}"'.format(master_pub))
+                master_pub.unlink()
+            except (FileNotFoundError, PermissionError):
+                pass
 
     if master_host:
         settings.setdefault('master_vagrant_ip', settings['vagrant_prefix'] + '.2.2')
